adjust refreshResultFieldClone
