+++ main/nextpatchnumber.txt
@@ -1,1 +1,1 @@
-98
+106
