+++ main/src/js/preparePage.js
@@ -12,8 +12,8 @@
  * if isEditor: prepareEditorApplet from editor.js
  */
 import $ from "jquery";
-import Hammer from "@egjs/hammerjs";
 import MQ from "./lib/mathquillWrapper.js";
+import Hammer from "@egjs/hammerjs";
 import mathQuillEditHandler, {
   mathQuillEditHandlerActive
 } from "./editHandler.js";
@@ -127,7 +127,7 @@ export async function mathQuillify(id) {
     temp = temp.replace(/\\mathrm/g, '');
     temp = temp.replace(/\\unit{/g, config.unit_replacement);
     temp = temp.replace(/\\cdot/g, config.multiplicationSign);
-    //TODO
+    //TODO simplify code 
     if (isEditor && isH5P()) {
       console.log('H5P & Editor');
       var mf = document.getElementById('math-field');
@@ -331,7 +331,8 @@ function clickHandler(ev) {
           if (cmd === '#Enter') {
             // mathQuillEditHandler cannot be outsourced to virtualKeyboard (circular dependency)
             fApp = FAList[_evnt.currentTarget.id];
-            // mathQuillEditHandler(fApp, MQ, 'enter'); //TODO special syntax?
+            //TODO see, if special syntax necessary? 
+            //TODO mathQuillEditHandler(fApp, MQ, 'enter'); 
             mathQuillEditHandler(fApp, MQ);
           } else {
             virtualKeyboardEventHandler(_evnt, cmd, fApp.mathField);
@@ -395,8 +396,7 @@ function unifyDefinitions(def) {
   }
   return dsList;
 }
-
-
+//TODO delete refreshLatex, replace with refreshTimesGlyph
 // function refreshLatex(lang) {
 //   var id;
 //   for (id in FAList) {
@@ -439,68 +439,3 @@ function unifyDefinitions(def) {
 //     }
 //   }
 // }
-
-
-async function refreshResultFieldClone(latex, fApp) {
-  console.log("refreshResultFieldClone");
-  latex = latex.replaceAll(config.unit_replacement, '\\unit{');
-  var parts = separateInputfield(latex);
-  var tex = parts.before + '{{result}}' + parts.after;
-  fApp.solution = parts.tag; //not encoded; used by formulaapplet-editor
-  // see updateTexinputEventHandler
-  var enc = encode(parts.tag);
-  console.log(tex + ' enc=' + enc + ' -> ' + decode(enc));
-  // latexHandler(tex, enc);
-  // $(document).trigger('texevent');
-
-  // H5P editor: send tex to formulaapplet-editor using dispatchEvent
-  if (isH5P()) {
-    var texinput = $('div.field.field-name-TEX_expression.text input')[0];
-    if (typeof texinput !== 'undefined') {
-      // value of TEX_expression field is set to EditorResult
-      texinput.value = tex;
-      // trigger InputEvent. 
-      // EventListener: updateTexinputEventHandler see formulaapplet-editor.js
-      texinput.dispatchEvent(new InputEvent('input', {
-        bubbles: true
-      }))
-    }
-  }
-  // getHTML
-  var html = '<p class="formula_applet" id="' + fApp.id;
-  if (fApp.hasSolution) {
-    html += '" data-b64="' + enc;
-  }
-  if (fApp.unitAuto) {
-    html += '" mode="physics';
-  }
-  html += '">' + tex + '</p>';
-  console.log(html);
-  var out = $('textarea#html_output');
-  if (typeof out !== 'undefined') {
-    out.text(html);
-  }
-}
-
-export function mathQuillifyEditor(fApp) {
-  // make whole mathFieldSpan editable
-  var mathFieldSpan = document.getElementById('math-field');
-  if (!mathFieldSpan) throw new Error("Cannot find math-field. The math editor must provide one.");
-  var editorMf = MQ.MathField(mathFieldSpan, {
-    spaceBehavesLikeTab: true, // configurable
-    handlers: {
-      edit: function (mathField) { // useful event handlers
-        try {
-          if (mathQuillEditHandlerActive.flag) {
-            var latex = mathField.latex();
-            console.log('mathQuillEditHandler refreshResultFieldClone latex=' + latex);
-            refreshResultFieldClone(latex, fApp);
-          }
-        } catch (error) {
-          console.error('ERROR in MQ.MathField: ' + error);
-        }
-      }
-    }
-  });
-  return editorMf;
-}
